refactor(serializers): apply patches from addon_serializer
This takes the patches being applied in addon_serializer.py and puts them directly in the
render window serializer. We should verify that there are no issues. But I did notice some
discrepancies:

1. I saw no difference in `extractRequiredFields()`
2. The addon serializer did not call `registerInstanceSerializer()` on `vtkStructuredPoints` with
the modified `imagedataSerializer` (only difference is that extent is used instead of dimensions).
3. The addon serializer did not call `registerInstanceSerializer()` on `vtkColorTransferFunction` with
the modified `colorTransferFunctionSerializer`.
4. `genericMapperSerializer()` only had debug message modifications
